move blueprint tests into test_app.py
